+++ o8g/Scripts/constants.py
@@ -65,9 +65,11 @@ regexHooks = dict( # A dictionary which holds the regex that then trigger each c
                   SetVarX =            re.compile(r'(?<![<,+-])SetVar'))
 
 specialHostPlacementAlgs = { # A Dictionary which holds tuples of X and Y placement offsets, for cards which place their hosted cards differently to normal, such as Personal Workshop
-                              'Personal Workshop' :            (-32,0),
-                              'The Supplier'      :            (-32,0),
-                              'Awakening Center'  :            (-32,0)
+                              'Personal Workshop'     :            (-32,0),
+                              'The Supplier'          :            (-32,0),
+                              'Awakening Center'      :            (-32,0),
+                              'London Library'        :            (-32,0),
+                              'Off-Campus Apartment'  :            (-32,0)
                            }